added different components for List, Update, Add ---functionality not implemented yet
