+++ src/redux/actions/songActionTypes.js
@@ -1,6 +1,7 @@
 export const FETCH_SONGS_REQUEST = 'FETCH_SONGS_REQUEST';
 export const FETCH_SONGS_SUCCESS = 'FETCH_SONGS_SUCCESS';
 export const FETCH_SONGS_FAILURE = 'FETCH_SONGS_FAILURE';
+export const UPDATE_SONG_SUCCESS = 'UPDATE_SONG_SUCCESS';
+export const UPDATE_SONG_FAILURE = 'UPDATE_SONG_FAILURE';
 export const ADD_SONG = 'ADD_SONG';
 export const DELETE_SONG = 'DELETE_SONG';
-export const UPDATE_SONG = 'UPDATE_SONG';
