+++ src/redux/actions/songActions.js
@@ -14,6 +14,16 @@ export const fetchSongsFailure = (error) => ({
   payload: error,
 });
 
+export const updateSongSuccess = (updatedSong) => ({
+  type: actionTypes.UPDATE_SONG_SUCCESS,
+  payload: updatedSong,
+});
+
+export const updateSongFailure = (error) => ({
+  type: actionTypes.UPDATE_SONG_FAILURE,
+  payload: error,
+});
+
 export const addSong = (song) => ({
   type: actionTypes.ADD_SONG,
   payload: song,
@@ -24,7 +34,17 @@ export const deleteSong = (id) => ({
   payload: id,
 });
 
-export const updateSong = (song) => ({
-  type: actionTypes.UPDATE_SONG,
-  payload: song,
-});
+export const updateSong = (songId, songData) => {
+  return (dispatch) => {
+    fetch(`https://my-json-server.typicode.com/Amansolo9/Mock-Json/musics${songId}`, {
+      method: 'PUT',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(songData),
+    })
+    .then((response) => response.json())
+    .then((data) => dispatch(updateSongSuccess(data)))
+    .catch((error) => dispatch(updateSongFailure(error.message)));
+  };
+};
